locust_standby: New locust; keeps a "standby" connection to the ensemble
The locust runner is used to ramp up a (presumably large) number of
Kazoo clients kept in "standby" mode (connected and processing
keep-alives, but otherwise inactive).

There is an "ensemble handler" which shuts down the leader 60s after
ramp-up is complete.

The collected metrics can be used to measure failover time.